Now keeping particle fade alpha and scale separate
Fixes issue with PDF_VELOCITYFADE, PDF_LIFEFADE and PDF_LIFESCALE stomping any changes made to the scale or transparency during the particles lifetime

(Also fixed some float -> double conversion warnings)

+++ src/playsim/p_actionfunctions.cpp
@@ -1783,7 +1783,7 @@ DEFINE_ACTION_FUNCTION(AActor, A_SpawnDefinedParticle)
 
 		if (particle)
 		{
-			particle->angle = angle.Degrees();
+			particle->angle = (float)angle.Degrees();
 		}
 	}
